:recycle: creates seperate client-side component
